+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - PR #836 Add ingest support for arrow chunked arrays in Column, Series, DataFrame creation
 - PR #763 Format doxygen comments for csv_read_arg struct
 - PR #532 CSV Reader: Use type dispatcher instead of switch block
+- PR #890 Type dispatcher: Support functors implemented for some, not all, column types
 
 ## Bug Fixes